+++ components/NavbarButton.tsx
@@ -21,7 +21,7 @@ export function NavbarButton({ color, icon, title, href, onClick, expanded }: Na
   const isExternal = !!href?.startsWith('http');
   const Icon = onClick ? IconChevronRight : icon;
   const pathname = usePathname();
-  // this is to make sure for works for trailing slashes
+  // make sure it works with trailing slashes
   const isCurrent = pathname === `${href}/` || pathname === href;
 
   const content = (
@@ -33,7 +33,7 @@ export function NavbarButton({ color, icon, title, href, onClick, expanded }: Na
         '--navbar-button-hover-bg-color-dark': isCurrent ? rgba(colors[color][6], 0.3) : rgba(colors.gray[6], 0.2),
       })}
       className={classes.root}
-      component="a"
+      component={isExternal ? 'a' : undefined}
       role={href ? undefined : 'button'}
       href={isExternal ? href : undefined}
       target={isExternal ? '_blank' : undefined}
@@ -57,7 +57,7 @@ export function NavbarButton({ color, icon, title, href, onClick, expanded }: Na
   return isExternal || onClick ? (
     content
   ) : (
-    <Link href={href as Route} passHref legacyBehavior>
+    <Link className={classes.link} href={href as Route}>
       {content}
     </Link>
   );