+++ package/DataTableRowSelectorCell.tsx
@@ -1,4 +1,5 @@
-import { Checkbox, MantineStyleProp, TableTd, type CheckboxProps } from '@mantine/core';
+import type { MantineStyleProp } from '@mantine/core';
+import { Checkbox, TableTd, type CheckboxProps } from '@mantine/core';
 import clsx from 'clsx';
 import type { DataTableSelectionTrigger } from './types';
 import { POINTER_CURSOR } from './utilityClasses';